Refactor progress-bar config
